# Slasher: Optimize span marshal and unmarshal