**What type of PR is this?**
Optimization

**What does this PR do? Why is it needed?**
This PR is a rough attempt at optimizing these often used functions by making sure they can inline and don't take up additional memory, results below:
```
BenchmarkSpan_Marshal_Old-8         29820471                37.1 ns/op            16 B/op          2 allocs/op
BenchmarkSpan_Unmarshal_New-8       137652634                9.81 ns/op            0 B/op          0 allocs/op
```
```
BenchmarkSpan_Marshal_New-8         80018118                15.3 ns/op             8 B/op          1 allocs/op
BenchmarkSpan_Unmarshal_New-8       171897799                6.51 ns/op            0 B/op          0 allocs/op
``` 

These changes could be controversial so let me know if these changes shouldn't be considered.